Postgres-python-fastapi/requirements.txt
psycopg2>=2.9.6,<3.0.0
python-dotenv>=1.0.1,<2.0.0
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.23.0,<0.24.0
pytest>=7.4.0,<9.0.0
httpx>=0.24.1,<0.30.0
